requirements-dev.txt
# Testing
pytest==6.2.2
pytest-cov==2.11.1
pytest-docker-tools==3.1.0
pytest-mock==3.6.1

# Data Wrangling
pandas==1.1.4

# Typehints
mypy==0.910

# Lints
flake8==3.8.4

# Black formatting
black==20.8b1

# Security Check
bandit==1.7.0

# Static Type Check
mypy==0.910

# Airflow
apache-airflow==2.1.0
apache-airflow-providers-postgres==2.0.0
apache-airflow-providers-mysql==2.0.0

# Kafka Python
kafka-python>=1.4.4

# MySQL
mysql-connector-python==8.0.11
mysqlclient==2.0.3